Bumped ffmpeg to 4.4

+++ lib-src/ffmpeg/libavutil/ripemd.h
@@ -19,9 +19,16 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
+/**
+ * @file
+ * @ingroup lavu_ripemd
+ * Public header for RIPEMD hash function implementation.
+ */
+
 #ifndef AVUTIL_RIPEMD_H
 #define AVUTIL_RIPEMD_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 #include "attributes.h"
@@ -29,7 +36,9 @@
 
 /**
  * @defgroup lavu_ripemd RIPEMD
- * @ingroup lavu_crypto
+ * @ingroup lavu_hash
+ * RIPEMD hash function implementation.
+ *
  * @{
  */
 
@@ -58,7 +67,7 @@ int av_ripemd_init(struct AVRIPEMD* context, int bits);
  * @param data    input data to update hash with
  * @param len     input data length
  */
-void av_ripemd_update(struct AVRIPEMD* context, const uint8_t* data, unsigned int len);
+void av_ripemd_update(struct AVRIPEMD* context, const uint8_t* data, size_t len);
 
 /**
  * Finish hashing and output digest value.
